Introduce buffered logger
Logging represents a significant overhead slowing down the computation
phase of the program. This commit introduces an optional buffered logger
that postpones results reporting to the end. Pass `-buffer=true` to use
it.

Optionally, you can pass `-cpuprofile=path` to generate a CPU profile of
the execution. It can be later inspected using the graphs that `go tool pprof
--addr=:12345 path` will show in the browser.